+++ src/instruction/ALU/ADD.h
@@ -33,14 +33,14 @@ namespace emulator{
             //Z = R ==0
             //C = (RD7&&RR7) || (RR7&&!R7) || (!R7 && RD7)
 
-            uint8_t R = at.memory.GP(RdVal)+at.memory.GP(RrVal);
+            int8_t R = (int8_t)at.memory.GP(RdVal)+(int8_t)at.memory.GP(RrVal);
 
             bool
                     RD3 =utils::isSet<3>(at.memory.GP(RdVal)),
                     RR3 =utils::isSet<3>(at.memory.GP(RrVal)),
                     R3 =utils::isSet<3>(R),
-                    RD7 =utils::isSet<7>(at.memory.GP(RdVal)),
-                    RR7 =utils::isSet<7>(at.memory.GP(RrVal)),
+                    RD7 =utils::isSet<7>((int8_t)at.memory.GP(RdVal)),
+                    RR7 =utils::isSet<7>((int8_t)at.memory.GP(RrVal)),
                     R7  =utils::isSet<7>(R);
             bool N=R7,
                  V=(RD7&&RR7&&!R7) || (!RD7&&!RR7&&R7);
